+++ src/main/java/at/sparklingscience/urbantrees/mapper/UiMapper.java
@@ -28,8 +28,8 @@ public interface UiMapper {
 	 * @return {@link Image}
 	 */
 	Image findImageForPhenologyObservationResult(
-			@Param("treeSpeciesId") int treeSpeciesId,
-			@Param("resultId") int resultId
+			@Param("treeSpeciesId") long treeSpeciesId,
+			@Param("resultId") long resultId
 			);
 	
 	/**
